refactor(terminal): extract TerminalResizePassScheduler
- Move the multi-terminal batch/chunked resize orchestration (batchResize,
  scheduleBatchResize, runResizePass, cancelActiveResizePass,
  executeResizePass, orderFocusedFirst) into a new sibling controller
- TerminalInstanceService keeps thin public delegates for
  scheduleBatchResize/runResizePass/cancelActiveResizePass (reached
  externally via the grid hook, AppLayout, and PaintFabricCompositor) and
  delegates dispose() cleanup
- Update the batchResize test file to reach the moved private method
  through resizePassScheduler
